+++ Modules/Intent.Modules.AspNetCore.IntegrationTests.CRUD/FactoryExtensions/TestImplementations/CrudMapHelper.cs
@@ -132,6 +132,8 @@ namespace Intent.Modules.AspNetCore.IntegrationTests.CRUD.FactoryExtensions.Test
             var getByIdOperation = operations.FirstOrDefault(o => string.Compare(o.Name, $"{getPrefix}{entityName}ById", ignoreCase: true) == 0 && ExpectedGetByIdParameters(o, entity, owningAggregate));
             if (getByIdOperation == null) return null;
 
+            var domainOperationInvocations = GetDomainInvocationOperations(operations, entity, owningAggregate);
+
             return new CrudMap(
                 testableProxy,
                 entity,
@@ -141,10 +143,37 @@ namespace Intent.Modules.AspNetCore.IntegrationTests.CRUD.FactoryExtensions.Test
                 operations.FirstOrDefault(o => string.Compare(o.Name, $"Delete{entityName}", ignoreCase: true) == 0 && ExpectedDeleteParameters(o, entity, owningAggregate)),
                 getByIdOperation,
                 operations.FirstOrDefault(o => string.Compare(o.Name, $"{getPrefix}{entityName.Pluralize(false)}", ignoreCase: true) == 0 && ExpectedGetAllParameters(o, owningAggregate)),
+                domainOperationInvocations,
                 responseDtoIdField,
                 owningAggregate
                 );
         }
+
+        private static IEnumerable<IHttpEndpointModel> GetDomainInvocationOperations(IEnumerable<IHttpEndpointModel> operations, ClassModel entity, ClassModel? owningAggregate)
+        {
+            return operations.Where(o => o.InternalElement.AssociatedElements.Any(ae => ae.Association.SpecializationTypeId == "9ea0382a-4617-412a-a8c8-af987bbce226"/*Update Entity Action*/  && ae.Association.SourceEnd?.ParentElement?.ParentElement?.Id == entity.Id)
+                                        && o.InternalElement.MappedToElements.Any(me => me.MappingTypeId == "d30bdba1-9c47-4917-b81d-29230fed5d6a"/* Method Invocation*/)
+                                        && ExpectedDomainInvocationOperationParameters(o, entity, owningAggregate)
+                                        );
+        }
+
+        private static bool ExpectedDomainInvocationOperationParameters(IHttpEndpointModel o, ClassModel entity, ClassModel? owningAggregate)
+        {
+            //This is backwards compatible with old composite rest urls
+            if (o.Inputs.Count == 2 &&
+                o.Inputs.First().TypeReference?.Element?.Id == entity.Attributes.FirstOrDefault(a => a.HasStereotype("Primary Key"))?.TypeReference?.Element?.Id)
+            {
+                return true;
+            }
+            if (owningAggregate != null)
+            {
+                return o.Inputs.Count == 3 &&
+                    o.Inputs.First().TypeReference?.Element?.Id == owningAggregate.Attributes.FirstOrDefault(a => a.HasStereotype("Primary Key"))?.TypeReference?.Element?.Id &&
+                    o.Inputs.ElementAt(1).TypeReference?.Element?.Id == entity.Attributes.FirstOrDefault(a => a.HasStereotype("Primary Key"))?.TypeReference?.Element?.Id;
+            }
+            return false;
+        }
+
 
         private static bool ExpectedGetAllParameters(IHttpEndpointModel o, ClassModel? owningAggregate)
         {